refactor(solver): split constraint walker helpers
