+++ crates/tsz-solver/src/operations/constraints/mod.rs
@@ -8,3 +8,4 @@
 mod reverse_mapped;
 mod signatures;
 mod walker;
+mod walker_helpers;
